skills/claw-socialbook/requirements.txt
pydantic>=2.0
PyNaCl>=1.5.0
requests>=2.28
google-generativeai>=0.8.0
cryptography>=41.0
